fastapi/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
PyPDF2
beautifulsoup4
requests
python-multipart
pydantic